+++ src/examples/src/config.ts
@@ -83,6 +83,7 @@ import BasicToolbar from './widgets/toolbar/Basic';
 import BasicTooltip from './widgets/tooltip/Basic';
 import ClickTooltip from './widgets/tooltip/Click';
 import FocusTooltip from './widgets/tooltip/Focus';
+import ListBox from './widgets/menu/ListBox';
 
 export interface ExampleConfig {
 	title?: string;
@@ -105,10 +106,12 @@ export interface Config {
 
 export const config: Config = {
 	'accordion-pane': {
-		examples: [{
-			filename: 'Exclusive',
-			module: Exclusive
-		}],
+		examples: [
+			{
+				filename: 'Exclusive',
+				module: Exclusive
+			}
+		],
 		filename: 'index',
 		overview: {
 			example: {
@@ -118,15 +121,18 @@ export const config: Config = {
 		}
 	},
 	button: {
-		examples: [{
-			filename: 'DisabledSubmit',
-			module: DisabledSubmit,
-			title: 'Disabled Submit Button'
-		}, {
-			filename: 'ToggleButton',
-			module: ToggleButton,
-			title: 'Toggle Button'
-		}],
+		examples: [
+			{
+				filename: 'DisabledSubmit',
+				module: DisabledSubmit,
+				title: 'Disabled Submit Button'
+			},
+			{
+				filename: 'ToggleButton',
+				module: ToggleButton,
+				title: 'Toggle Button'
+			}
+		],
 		filename: 'index',
 		overview: {
 			example: {
@@ -136,16 +142,19 @@ export const config: Config = {
 		}
 	},
 	calendar: {
-		examples: [{
-			filename: 'CustomFirstWeekDay',
-			module: FirstDayOfWeekCalendar,
-			title: 'Custom First Day of Week'
-		}, {
-			description: 'Demonstrates limiting the selectable region of the calendar.',
-			filename: 'LimitedRange',
-			module: LimitedRange,
-			title: 'Limited Date Range'
-		}],
+		examples: [
+			{
+				filename: 'CustomFirstWeekDay',
+				module: FirstDayOfWeekCalendar,
+				title: 'Custom First Day of Week'
+			},
+			{
+				description: 'Demonstrates limiting the selectable region of the calendar.',
+				filename: 'LimitedRange',
+				module: LimitedRange,
+				title: 'Limited Date Range'
+			}
+		],
 		filename: 'index',
 		overview: {
 			example: {
@@ -173,19 +182,23 @@ export const config: Config = {
 		}
 	},
 	'checkbox-group': {
-		examples: [{
-			filename: 'InitialValue',
-			module: InitialValueCheckboxGroup,
-			title: 'Initial Value'
-		}, {
-			filename: 'CustomLabel',
-			module: CustomLabelCheckboxGroup,
-			title: 'Custom Label'
-		}, {
-			filename: 'CustomRenderer',
-			module: CustomRendererCheckboxGroup,
-			title: 'Custom Renderer'
-		}],
+		examples: [
+			{
+				filename: 'InitialValue',
+				module: InitialValueCheckboxGroup,
+				title: 'Initial Value'
+			},
+			{
+				filename: 'CustomLabel',
+				module: CustomLabelCheckboxGroup,
+				title: 'Custom Label'
+			},
+			{
+				filename: 'CustomRenderer',
+				module: CustomRendererCheckboxGroup,
+				title: 'Custom Renderer'
+			}
+		],
 		filename: 'index',
 		overview: {
 			example: {
@@ -195,31 +208,38 @@ export const config: Config = {
 		}
 	},
 	chip: {
-		examples: [{
-			filename: 'Icon',
-			module: IconChip,
-			title: 'Icon'
-		}, {
-			filename: 'Clickable',
-			module: ClickableChip,
-			title: 'Clickable'
-		}, {
-			filename: 'Disabled',
-			module: DisabledChip,
-			title: 'Disabled'
-		}, {
-			filename: 'Closable',
-			module: ClosableChip,
-			title: 'Closable'
-		}, {
-			filename: 'ClosableRenderer',
-			module: ClosableRendererChip,
-			title: 'Closable Renderer'
-		}, {
-			filename: 'ClickableClosable',
-			module: ClickableClosableChip,
-			title: 'Clickable and Closable'
-		}],
+		examples: [
+			{
+				filename: 'Icon',
+				module: IconChip,
+				title: 'Icon'
+			},
+			{
+				filename: 'Clickable',
+				module: ClickableChip,
+				title: 'Clickable'
+			},
+			{
+				filename: 'Disabled',
+				module: DisabledChip,
+				title: 'Disabled'
+			},
+			{
+				filename: 'Closable',
+				module: ClosableChip,
+				title: 'Closable'
+			},
+			{
+				filename: 'ClosableRenderer',
+				module: ClosableRendererChip,
+				title: 'Closable Renderer'
+			},
+			{
+				filename: 'ClickableClosable',
+				module: ClickableClosableChip,
+				title: 'Clickable and Closable'
+			}
+		],
 		filename: 'index',
 		overview: {
 			example: {
@@ -237,11 +257,13 @@ export const config: Config = {
 		}
 	},
 	'constrained-input': {
-		examples: [{
-			filename: 'Username',
-			module: Username,
-			title: 'Username'
-		}],
+		examples: [
+			{
+				filename: 'Username',
+				module: Username,
+				title: 'Username'
+			}
+		],
 		overview: {
 			example: {
 				filename: 'Basic',
@@ -267,36 +289,44 @@ export const config: Config = {
 		}
 	},
 	grid: {
-		examples: [{
-			filename: 'Sorting',
-			module: Sorting,
-			title: 'Grid with Sorting'
-		}, {
-			filename: 'Filtering',
-			module: Filtering,
-			title: 'Grid with Filtered Columns'
-		}, {
-			filename: 'CustomCellRenderer',
-			module: CustomCellRenderer,
-			title: 'Grid with Custom Cell Rendering'
-		}, {
-			filename: 'CustomFilterRenderer',
-			module: GridCustomFilterRenderer,
-			title: 'Grid with Custom Filter Renderer'
-		}, {
-			filename: 'CustomSortRenderer',
-			module: CustomSortRenderer,
-			title: 'Grid with Customized Sort Rendering'
-		}, {
-			filename: 'Restful',
-			module: Restful,
-			title: 'Grid with Data Loaded from a Remote Resource (REST)'
-		}, {
-			description: 'Demonstrates using the grid utilities to support editable cells',
-			filename: 'EditableCells',
-			module: EditableCells,
-			title: 'Grid with Editable Cells'
-		}],
+		examples: [
+			{
+				filename: 'Sorting',
+				module: Sorting,
+				title: 'Grid with Sorting'
+			},
+			{
+				filename: 'Filtering',
+				module: Filtering,
+				title: 'Grid with Filtered Columns'
+			},
+			{
+				filename: 'CustomCellRenderer',
+				module: CustomCellRenderer,
+				title: 'Grid with Custom Cell Rendering'
+			},
+			{
+				filename: 'CustomFilterRenderer',
+				module: GridCustomFilterRenderer,
+				title: 'Grid with Custom Filter Renderer'
+			},
+			{
+				filename: 'CustomSortRenderer',
+				module: CustomSortRenderer,
+				title: 'Grid with Customized Sort Rendering'
+			},
+			{
+				filename: 'Restful',
+				module: Restful,
+				title: 'Grid with Data Loaded from a Remote Resource (REST)'
+			},
+			{
+				description: 'Demonstrates using the grid utilities to support editable cells',
+				filename: 'EditableCells',
+				module: EditableCells,
+				title: 'Grid with Editable Cells'
+			}
+		],
 		filename: 'index',
 		overview: {
 			example: {
@@ -306,11 +336,13 @@ export const config: Config = {
 		}
 	},
 	icon: {
-		examples: [{
-			filename: 'AltText',
-			module: AltTextIcon,
-			title: 'With alt text'
-		}],
+		examples: [
+			{
+				filename: 'AltText',
+				module: AltTextIcon,
+				title: 'With alt text'
+			}
+		],
 		overview: {
 			example: {
 				filename: 'Basic',
@@ -335,16 +367,27 @@ export const config: Config = {
 		}
 	},
 	menu: {
-		examples: [{
-			description: 'In this example, an activeIndex and onActiveIndexChange property are passed into the Menu allowing for thr active menu item to be controlled.',
-			filename: 'Controlled',
-			module: ControlledMenu,
-			title: 'Controlled Active Index'
-		}, {
-			filename: 'ItemRenderer',
-			module: ItemRenderer,
-			title: 'Item Renderer'
-		}],
+		examples: [
+			{
+				description:
+					'In this example, an activeIndex and onActiveIndexChange property are passed into the Menu allowing for thr active menu item to be controlled.',
+				filename: 'Controlled',
+				module: ControlledMenu,
+				title: 'Controlled Active Index'
+			},
+			{
+				filename: 'ItemRenderer',
+				module: ItemRenderer,
+				title: 'Item Renderer'
+			},
+			{
+				description:
+					'This example shows the menu used as a Listbox. This allows for a selection to be made and persisted. Useful for user selections and within selects / typeahead etc.',
+				filename: 'ListBox',
+				module: ListBox,
+				title: 'List Box'
+			}
+		],
 		overview: {
 			example: {
 				filename: 'Basic',
@@ -401,15 +444,18 @@ export const config: Config = {
 		}
 	},
 	select: {
-		examples: [{
-			filename: 'AdvancedOptions',
-			module: AdvancedOptions,
-			title: 'Advanced options'
-		}, {
-			filename: 'NonNative',
-			module: NonNative,
-			title: 'Non Native'
-		}],
+		examples: [
+			{
+				filename: 'AdvancedOptions',
+				module: AdvancedOptions,
+				title: 'Advanced options'
+			},
+			{
+				filename: 'NonNative',
+				module: NonNative,
+				title: 'Non Native'
+			}
+		],
 		overview: {
 			example: {
 				filename: 'Basic',
@@ -450,11 +496,13 @@ export const config: Config = {
 		}
 	},
 	'tab-controller': {
-		examples: [{
-			filename: 'Disabled',
-			module: DisabledTabController,
-			title: 'TabController with disabled tabs'
-		}],
+		examples: [
+			{
+				filename: 'Disabled',
+				module: DisabledTabController,
+				title: 'TabController with disabled tabs'
+			}
+		],
 		filename: 'index',
 		overview: {
 			example: {
@@ -464,27 +512,33 @@ export const config: Config = {
 		}
 	},
 	'text-area': {
-		examples: [{
-			filename: 'Disabled',
-			module: DisabledTextArea,
-			title: 'Disabled'
-		}, {
-			filename: 'HelperText',
-			module: HelperTextTextArea,
-			title: 'Helper text'
-		}, {
-			filename: 'HiddenLabel',
-			module: HiddenLabelTextArea,
-			title: 'Hidden label'
-		}, {
-			filename: 'ValidatedCustom',
-			module: ValidatedCustomTextArea,
-			title: 'Validated with custom validator'
-		}, {
-			filename: 'ValidatedRequired',
-			module: ValidatedRequiredTextArea,
-			title: 'Validated required'
-		}],
+		examples: [
+			{
+				filename: 'Disabled',
+				module: DisabledTextArea,
+				title: 'Disabled'
+			},
+			{
+				filename: 'HelperText',
+				module: HelperTextTextArea,
+				title: 'Helper text'
+			},
+			{
+				filename: 'HiddenLabel',
+				module: HiddenLabelTextArea,
+				title: 'Hidden label'
+			},
+			{
+				filename: 'ValidatedCustom',
+				module: ValidatedCustomTextArea,
+				title: 'Validated with custom validator'
+			},
+			{
+				filename: 'ValidatedRequired',
+				module: ValidatedRequiredTextArea,
+				title: 'Validated required'
+			}
+		],
 		overview: {
 			example: {
 				filename: 'Basic',
@@ -493,31 +547,38 @@ export const config: Config = {
 		}
 	},
 	'text-input': {
-		examples: [{
-			filename: 'WithLabel',
-			module: TextInputWithLabel,
-			title: 'TextInput with Label'
-		}, {
-			filename: 'HiddenLabel',
-			module: HiddenLabelTextInput,
-			title: 'TextInput with hidden label'
-		}, {
-			filename: 'Disabled',
-			module: DisabledTextInput,
-			title: 'Disabled TextInput'
-		}, {
-			filename: 'Validated',
-			module: ValidatedTextInput,
-			title: 'Validated TextInput'
-		}, {
-			filename: 'HelperText',
-			module: HelperTextInput,
-			title: 'TextInput with helper text'
-		}, {
-			filename: 'LeadingTrailing',
-			module: LeadingTrailingTextInput,
-			title: 'TextInput with leading and trailing'
-		}],
+		examples: [
+			{
+				filename: 'WithLabel',
+				module: TextInputWithLabel,
+				title: 'TextInput with Label'
+			},
+			{
+				filename: 'HiddenLabel',
+				module: HiddenLabelTextInput,
+				title: 'TextInput with hidden label'
+			},
+			{
+				filename: 'Disabled',
+				module: DisabledTextInput,
+				title: 'Disabled TextInput'
+			},
+			{
+				filename: 'Validated',
+				module: ValidatedTextInput,
+				title: 'Validated TextInput'
+			},
+			{
+				filename: 'HelperText',
+				module: HelperTextInput,
+				title: 'TextInput with helper text'
+			},
+			{
+				filename: 'LeadingTrailing',
+				module: LeadingTrailingTextInput,
+				title: 'TextInput with leading and trailing'
+			}
+		],
 		filename: 'index',
 		overview: {
 			example: {
@@ -527,39 +588,48 @@ export const config: Config = {
 		}
 	},
 	'time-picker': {
-		examples: [{
-			filename: 'FilteredOnInput',
-			module: FilteredOnInputTimePicker,
-			title: 'Filtered on input'
-		}, {
-			filename: 'OpensOnFocus',
-			module: OpenOnFocusTimePicker,
-			title: 'Opens on focus'
-		}, {
-			filename: 'DisabledMenuItems',
-			module: DisabledMenuItemsTimePicker,
-			title: 'Disabled menu items'
-		}, {
-			filename: 'Disabled',
-			module: DisabledTimePicker,
-			title: 'Disabled time picker'
-		}, {
-			filename: 'SelectBySecond',
-			module: SelectBySecondTimePicker,
-			title: 'Select time by the second'
-		}, {
-			filename: '12HourTime',
-			module: TwelveHourTimePicker,
-			title: '12 hour time'
-		}, {
-			filename: 'Required',
-			module: RequiredTimePicker,
-			title: 'Required time picker'
-		}, {
-			filename: 'Native',
-			module: NativeTimePicker,
-			title: 'Native time picker'
-		}],
+		examples: [
+			{
+				filename: 'FilteredOnInput',
+				module: FilteredOnInputTimePicker,
+				title: 'Filtered on input'
+			},
+			{
+				filename: 'OpensOnFocus',
+				module: OpenOnFocusTimePicker,
+				title: 'Opens on focus'
+			},
+			{
+				filename: 'DisabledMenuItems',
+				module: DisabledMenuItemsTimePicker,
+				title: 'Disabled menu items'
+			},
+			{
+				filename: 'Disabled',
+				module: DisabledTimePicker,
+				title: 'Disabled time picker'
+			},
+			{
+				filename: 'SelectBySecond',
+				module: SelectBySecondTimePicker,
+				title: 'Select time by the second'
+			},
+			{
+				filename: '12HourTime',
+				module: TwelveHourTimePicker,
+				title: '12 hour time'
+			},
+			{
+				filename: 'Required',
+				module: RequiredTimePicker,
+				title: 'Required time picker'
+			},
+			{
+				filename: 'Native',
+				module: NativeTimePicker,
+				title: 'Native time picker'
+			}
+		],
 		overview: {
 			example: {
 				filename: 'Basic',
@@ -568,15 +638,18 @@ export const config: Config = {
 		}
 	},
 	'title-pane': {
-		examples: [{
-			filename: 'AriaHeadingLevel',
-			module: HeadingLevel,
-			title: 'Defined aria heading level'
-		}, {
-			filename: 'NonCloseable',
-			module: NonCloseable,
-			title: 'Non closeable'
-		}],
+		examples: [
+			{
+				filename: 'AriaHeadingLevel',
+				module: HeadingLevel,
+				title: 'Defined aria heading level'
+			},
+			{
+				filename: 'NonCloseable',
+				module: NonCloseable,
+				title: 'Non closeable'
+			}
+		],
 		overview: {
 			example: {
 				filename: 'Basic',
@@ -593,15 +666,18 @@ export const config: Config = {
 		}
 	},
 	tooltip: {
-		examples: [{
-			filename: 'Focus',
-			module: FocusTooltip,
-			title: 'Show on focus'
-		}, {
-			filename: 'Click',
-			module: ClickTooltip,
-			title: 'Show on click'
-		}],
+		examples: [
+			{
+				filename: 'Focus',
+				module: FocusTooltip,
+				title: 'Show on focus'
+			},
+			{
+				filename: 'Click',
+				module: ClickTooltip,
+				title: 'Show on click'
+			}
+		],
 		overview: {
 			example: {
 				filename: 'Basic',
