+++ src/examples/src/widgets/menu/ListBox.tsx
@@ -1,25 +1,22 @@
 import { create, tsx } from '@dojo/framework/core/vdom';
 import Menu from '@dojo/widgets/menu';
+import states from './states';
 import icache from '@dojo/framework/core/middleware/icache';
 
 const factory = create({ icache });
 
-export default factory(function Basic({ middleware: { icache } }) {
-	const options = [
-		{ value: 'Save' },
-		{ value: 'copy', label: 'Copy' },
-		{ value: 'Paste', disabled: true }
-	];
-
+export default factory(function ListBox({ middleware: { icache } }) {
 	return (
 		<virtual>
 			<Menu
-				options={options}
+				listBox
+				options={states}
 				onValue={(value) => {
 					icache.set('value', value);
 				}}
+				itemsInView={8}
 			/>
-			<p>{`Clicked on: ${icache.getOrSet('value', '')}`}</p>{' '}
+			<p>{`Selected: ${icache.getOrSet('value', '')}`}</p>
 		</virtual>
 	);
 });
