+++ src/theme/list-box-item.m.css
@@ -7,6 +7,11 @@
 	height: 45px;
 }
 
+/* Added to selected items */
+.selected {
+	background: grey;
+}
+
 /* Added to active items */
 .active {
 	border: 1px solid blue;
