+++ src/theme/list-box-item.m.css.d.ts
@@ -1,3 +1,4 @@
 export const root: string;
+export const selected: string;
 export const active: string;
 export const disabled: string;
